actions files in server dir
